+++ tests/__init__.py
@@ -1,23 +1,29 @@
 import base_test
+import test_abstract_grouping
+import test_colors
 import test_control
 import test_curve
 import test_dag_node
+import test_hierarchy_chain
 import test_joint
-import test_MVP
+import test_meta_data
 import test_rig
+import test_sub_rig
 import test_transform
-import test_build_biped
-import test_template_sub_rigs
 import test_unicode_delegate
 
-__all__ = [base_test,
-           test_control,
-           test_curve,
-           test_dag_node,
-           test_joint,
-           test_MVP,
-           test_rig,
-           test_transform,
-           test_build_biped,
-           test_template_sub_rigs,
-           test_unicode_delegate]
+__all__ = [
+    base_test,
+    test_abstract_grouping,
+    test_colors,
+    test_control,
+    test_curve,
+    test_dag_node,
+    test_hierarchy_chain,
+    test_joint,
+    test_meta_data,
+    test_rig,
+    test_sub_rig,
+    test_transform,
+    test_unicode_delegate,
+]
