+++ tests/acceptance/__init__.py
@@ -1,0 +1,11 @@
+import test_biped_foot
+import test_build_biped
+import test_MVP
+import test_biped_average_time
+
+__all__ = [
+    test_biped_average_time,
+    test_biped_foot,
+    test_build_biped,
+    test_MVP
+]
